Fixes for conda-build 1.20.3
Use the internal API directly. Ironically, the output issues were common
to many projects and so they have suppressed the output themselves. I
maintain the suppress contextmanager here just in case we need it in the
future. Also, add the flush method to the fake sink as that is expected
as part of a fake pipe.